Remove latest folder, move top level sources to common
