+++ typescript/src/common/Reader.ts
