+++ typescript/src/common/StreamBuffer.ts
