+++ typescript/src/common/getBigUint64.ts
